requirements.txt
langchain==0.3.14
langchain_cohere==0.3.4
langchain_community==0.3.14
langchain_core==0.3.29
langchain_groq==0.2.3
pydantic==2.10.5
python-dotenv==1.0.1
streamlit==1.39.0
streamlit_chat==0.1.1
